/channels を追加
