+++ src/main/kotlin/blue/starry/saya/models/Comments.kt
@@ -57,19 +57,6 @@ data class Comment(
 }
 
 @Serializable
-data class JikkyoChannel(
-    val type: Channel.Type,
-    val jk: Int? = null,
-    val name: String,
-    val serviceIds: Set<Int>,
-    val tags: Set<String> = emptySet(),
-    val isOfficial: Boolean = false,
-    val miyouId: String? = null,
-    val communities: Set<String> = emptySet(),
-    val hashtags: Set<String> = emptySet()
-)
-
-@Serializable
 data class TimeshiftCommentControl(
     val action: Action,
     val seconds: Double = 0.0
